refactor: api structure
