+++ src/api/http/profile.ts
@@ -1,6 +1,6 @@
 import { http } from './http';
-import { HeaderValue, UrlSubString } from '../utils/types/enums';
-import type { IPhotoData, IRequestProfile, IResponse, IResponseProfile } from '../utils/types/api';
+import { HeaderValue, UrlSubString } from '../../utils/types/enums';
+import type { IPhotoData, IRequestProfile, IResponse, IResponseProfile } from '../../utils/types/api';
 
 export const getProfileAPI = async (id: number): Promise<IResponseProfile> => {
   const response = await http.get(UrlSubString.profile + id);
